+++ src/eyepop/webrtc_whip.ts
@@ -1,12 +1,12 @@
 import {v4 as uuidv4} from 'uuid'
-import {LiveMedia, SessionPlus} from './types'
+import {LiveMedia, WorkerSession} from './types'
 import {HttpClient} from './shims/http_client'
 import {Logger} from 'pino'
 import {WebrtcBase} from './webrtc_base'
 
 export class WebrtcWhip extends WebrtcBase implements LiveMedia {
     private _stream: MediaStream | null
-    constructor(stream: MediaStream, getSession: () => Promise<SessionPlus>, client: HttpClient, requestLogger: Logger) {
+    constructor(stream: MediaStream, getSession: () => Promise<WorkerSession>, client: HttpClient, requestLogger: Logger) {
         super(getSession, client, uuidv4().toString(), '/liveIngress/whip', requestLogger)
         this._stream = stream
     }
